@@ -17,19 +17,19 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>3.3.10</version>
+        <version>3.4.7</version>
         <relativePath />
     </parent>
 
     <properties>
         <failOnDependencyWarning>true</failOnDependencyWarning>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>17</java.version>
-        <cxf.version>4.1.1</cxf.version>
-        <snakeyaml.version>2.3</snakeyaml.version>
+        <java.version>21</java.version>
+        <cxf.version>4.1.2</cxf.version>
+        <snakeyaml.version>2.4</snakeyaml.version>
         <lucene.version>9.12.0</lucene.version>
         <docker.container.repo>ghcr.io/kagemomiji/airsonic-advanced</docker.container.repo>
-        <docker.java.version>17.0.10_7</docker.java.version>
+        <docker.java.version>21.0.7_6</docker.java.version>
     </properties>
 
     <repositories>
@@ -88,7 +88,7 @@
                 <artifactId>asm</artifactId>
                 <version>9.7</version>
             </dependency>
-         <dependency>
+            <dependency>
                 <groupId>org.apache.commons</groupId>
                 <artifactId>commons-lang3</artifactId>
                 <version>3.16.0</version>
@@ -117,6 +117,11 @@
                 <groupId>jakarta.xml.bind</groupId>
                 <artifactId>jakarta.xml.bind-api</artifactId>
                 <version>4.0.2</version>
+            </dependency>
+            <dependency>
+                <groupId>javax.xml.bind</groupId>
+                <artifactId>jaxb-api</artifactId>
+                <version>2.3.1</version>
             </dependency>
             <dependency>
                 <groupId>org.glassfish.jaxb</groupId>
@@ -166,7 +171,7 @@
             <dependency>
                 <groupId>net.java.dev.jna</groupId>
                 <artifactId>jna</artifactId>
-                <version>5.15.0</version>
+                <version>5.17.0</version>
             </dependency>
             <dependency>
                 <groupId>net.java.dev.jna</groupId>
