@@ -36,8 +36,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.security.test.context.support.WithMockUser;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 
 import java.nio.file.Path;
@@ -65,10 +65,10 @@ public class PodcastEpisodesControllerTest {
     @TempDir
     private static Path tempDir;
 
-    @MockBean
+    @MockitoBean
     private PodcastPersistenceService podcastService;
 
-    @MockBean
+    @MockitoBean
     private PodcastDownloadClient podcastDownloadClient;
 
     @Mock
