@@ -33,8 +33,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -53,7 +53,7 @@ import static org.mockito.Mockito.when;
 @EnableConfigurationProperties({AirsonicHomeConfig.class})
 public class PodcastDownloadClientIntegrationTest {
 
-    @MockBean
+    @MockitoBean
     private VersionService versionService;
 
     @TempDir
