@@ -14,20 +14,18 @@ import java.util.Optional;
 @Repository
 public interface PodcastEpisodeRepository extends JpaRepository<PodcastEpisode, Integer> {
 
-    public Optional<PodcastEpisode> findByChannelAndUrl(PodcastChannel channel, String url);
+    Optional<PodcastEpisode> findByChannelAndUrl(PodcastChannel channel, String url);
 
-    public Optional<PodcastEpisode> findByChannelAndEpisodeGuid(PodcastChannel channel, String episodeGuid);
+    Optional<PodcastEpisode> findByChannelAndEpisodeGuid(PodcastChannel channel, String episodeGuid);
 
-    public Optional<PodcastEpisode> findByChannelAndTitleAndPublishDate(PodcastChannel channel, String title,
-            Instant publishDate);
+    Optional<PodcastEpisode> findByChannelAndTitleAndPublishDate(PodcastChannel channel, String title, Instant publishDate);
 
-    public List<PodcastEpisode> findByChannel(PodcastChannel channel);
+    List<PodcastEpisode> findByChannel(PodcastChannel channel);
 
-    public List<PodcastEpisode> findByChannelAndLockedFalse(PodcastChannel channel);
+    List<PodcastEpisode> findByChannelAndLockedFalse(PodcastChannel channel);
 
-    public List<PodcastEpisode> findByChannelAndStatus(PodcastChannel channel, PodcastStatus status);
+    List<PodcastEpisode> findByChannelAndStatus(PodcastChannel channel, PodcastStatus status);
 
-    public List<PodcastEpisode> findByStatusAndPublishDateNotNullAndMediaFilePresentTrueOrderByPublishDateDesc(
+    List<PodcastEpisode> findByStatusAndPublishDateNotNullAndMediaFilePresentTrueOrderByPublishDateDesc(
                     PodcastStatus status);
-
 }
